pyproject.toml
[tool.poetry]
name = "cellmap-api-aws"
version = "0.1.0"
description = "Deployment for the cellmap data api"
authors = ["Davis Vann Bennett <davis.v.bennett@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
aws-cdk-lib = "2.36.0"
constructs = "10.0.0,<11.0.0"

[tool.poetry.dev-dependencies]
mypy = "pytest"
pytest = "^7.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
